1.remove start commit from branch, 2.use repository manager of the main engine while loading new repo from XML, 3.add NoActiveBranchError if the user try to make some action while there is no active repository yet

+++ engine/src/MainEngine.java
@@ -1,22 +1,25 @@
 import exceptions.InvalidBranchNameError;
+import exceptions.NoActiveBranchError;
+import exceptions.NoActiveRepositoryError;
 import exceptions.UncommittedChangesError;
 
 import java.util.*;
 
 public class MainEngine {
 
-    private RepositoryManager repositoryManager;
+    private static RepositoryManager repositoryManager;
 
     public MainEngine(){
-        // TODO - remove before submit
-        String fullPath = "C:\\test";
-        Settings.setNewRepository(fullPath);
         repositoryManager = new RepositoryManager();
-        repositoryManager.switchActiveRepository(fullPath);
-//        repositoryManager.createNewRepository(fullPath);
     }
 
-    public Map<String ,List<String>> getWorkingCopyStatus(){
+    public static RepositoryManager getRepositoryManager(){
+        return repositoryManager;
+    }
+    public Map<String ,List<String>> getWorkingCopyStatus() throws NoActiveRepositoryError{
+        if(repositoryManager.getActiveRepository() == null){
+            throw new NoActiveRepositoryError("No active repository yet");
+        }
         return repositoryManager.getActiveRepository().getWorkingCopy();
     }
 
@@ -43,15 +46,29 @@ public class MainEngine {
         }
     }
 
-    public List<String> getCurrentCommitState(){
+    public List<String> getCurrentCommitState()
+            throws NoActiveRepositoryError, UncommittedChangesError{
+        if(repositoryManager.getActiveRepository() == null){
+            throw new NoActiveRepositoryError("No active repository yet");
+        }
+        if(repositoryManager.getActiveRepository().getActiveBranch().getHead() == null){
+            throw new UncommittedChangesError("No active commit");
+        }
         return repositoryManager.getActiveRepository().getActiveBranch().getCommittedState();
     }
 
-    public void createNewBranch(String name, boolean checkout) throws InvalidBranchNameError, UncommittedChangesError{
+    public void createNewBranch(String name, boolean checkout)
+            throws InvalidBranchNameError, UncommittedChangesError, NoActiveRepositoryError{
+        if(repositoryManager.getActiveRepository() == null){
+            throw new NoActiveRepositoryError("No active repository yet");
+        }
         repositoryManager.getActiveRepository().createNewBranch(name, checkout);
     }
 
-    public List<String> getAllBranches(){
+    public List<String> getAllBranches() throws NoActiveRepositoryError{
+        if(repositoryManager.getActiveRepository() == null){
+            throw new NoActiveRepositoryError("No active repository yet");
+        }
         List<String> allBranchesName = repositoryManager.getActiveRepository().getAllBranches();
         String headBranch = repositoryManager.getActiveRepository().getActiveBranch().getName();
         List<String> res = new ArrayList<>();
@@ -64,20 +81,32 @@ public class MainEngine {
         return res;
     }
 
-    public void deleteBranch(String name) throws InvalidBranchNameError, IllegalArgumentException{
+    public void deleteBranch(String name) throws InvalidBranchNameError, IllegalArgumentException, NoActiveRepositoryError{
+        if(repositoryManager.getActiveRepository() == null){
+            throw new NoActiveRepositoryError("No active repository yet");
+        }
         repositoryManager.getActiveRepository().deleteBranch(name);
     }
 
-    public void checkoutBranch(String name, boolean force) throws InvalidBranchNameError, UncommittedChangesError {
+    public void checkoutBranch(String name, boolean force) throws InvalidBranchNameError, UncommittedChangesError, NoActiveRepositoryError {
+        if(repositoryManager.getActiveRepository() == null){
+            throw new NoActiveRepositoryError("No active repository yet");
+        }
        repositoryManager.getActiveRepository().checkoutBranch(name, force);
     }
 
-    public List<String> getActiveBranchHistory(){
+    public List<String> getActiveBranchHistory() throws NoActiveRepositoryError{
+        if(repositoryManager.getActiveRepository() == null){
+            throw new NoActiveRepositoryError("No active repository yet");
+        }
         return repositoryManager.getActiveRepository().getActiveBranch().getCommitHistory();
     }
 
-    public String getCurrentRepoName() {
-        return repositoryManager.getActiveRepository().getName();
+    public String getCurrentRepoName() throws NoActiveRepositoryError{
+        if(repositoryManager.getActiveRepository() == null){
+            throw new NoActiveRepositoryError("No active repository yet");
+        }
+        return repositoryManager.getActiveRepository() != null ? repositoryManager.getActiveRepository().getName() : "";
     }
 
     public String isXmlValid(String xmlPath){
@@ -88,6 +117,9 @@ public class MainEngine {
             return null;
         } catch (XmlException e) {
             return e.message;
+        }
+        catch (UncommittedChangesError | InvalidBranchNameError e){
+            return e.getMessage();
         }
     }
